fix: flatten server-storage/telemetry handlers
