--- apps/server-storage/src/adapters/http/controllers/bucket-file/handlers/read.ts
@@ -1,116 +1,0 @@
-/*
- * Copyright (c) 2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import { useRequestQuery } from '@routup/basic/query';
-import type { Request, Response } from 'routup';
-import { send, useRequestParam } from 'routup';
-import {
-    applyQuery,
-} from 'typeorm-extension';
-import { NotFoundError } from '@ebec/http';
-import type { DataSource } from 'typeorm';
-import { BucketFileEntity } from '../../../../database/index.ts';
-
-export async function executeBucketFileRouteGetOneHandler(
-    req: Request,
-    res: Response,
-    dataSource: DataSource,
-) : Promise<any> {
-    const id = useRequestParam(req, 'id');
-
-    const repository = dataSource.getRepository(BucketFileEntity);
-    const query = repository.createQueryBuilder('bucketFile')
-        .where('bucketFile.id = :id', { id });
-
-    applyQuery(query, useRequestQuery(req), {
-        defaultAlias: 'bucketFile',
-        fields: {
-            default: [
-                'id',
-                'name',
-                'path',
-                'directory',
-                'size',
-                'hash',
-                'created_at',
-                'updated_at',
-                'realm_id',
-                'actor_type',
-                'actor_id',
-                'bucket_id',
-            ],
-        },
-        relations: { allowed: ['bucket'] },
-    });
-
-    const entity = await query.getOne();
-
-    if (!entity) {
-        throw new NotFoundError();
-    }
-
-    return send(res, entity);
-}
-
-export async function executeBucketFileRouteGetManyHandler(
-    req: Request,
-    res: Response,
-    dataSource: DataSource,
-) : Promise<any> {
-    const repository = dataSource.getRepository(BucketFileEntity);
-    const query = repository.createQueryBuilder('bucketFile');
-    query.groupBy('bucketFile.id');
-
-    const { pagination } = applyQuery(query, useRequestQuery(req), {
-        defaultAlias: 'bucketFile',
-        fields: {
-            default: [
-                'id',
-                'name',
-                'path',
-                'directory',
-                'size',
-                'hash',
-                'created_at',
-                'updated_at',
-                'realm_id',
-                'actor_type',
-                'actor_id',
-                'bucket_id',
-            ],
-        },
-        relations: {
-            allowed: ['bucket'],
-            onJoin: (_property, key, query) => {
-                query.addGroupBy(`${key}.id`);
-            },
-        },
-        filters: {
-            allowed: [
-                'id',
-                'name',
-                'directory',
-                'realm_id',
-                'actor_type',
-                'actor_id',
-                'bucket_id',
-            ],
-        },
-        pagination: { maxLimit: 50 },
-        sort: { allowed: ['id', 'directory', 'name', 'updated_at', 'created_at'] },
-    });
-
-    const [entities, total] = await query.getManyAndCount();
-
-    return send(res, {
-        data: entities,
-        meta: {
-            total,
-            ...pagination,
-        },
-    });
-}
