--- apps/server-storage/src/adapters/http/controllers/bucket/handlers/index.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './create.ts';
-export * from './delete.ts';
-export * from './read.ts';
-export * from './stream.ts';
-export * from './update.ts';
-export * from './upload.ts';
