--- apps/server-storage/src/adapters/http/controllers/bucket/handlers/read.ts
@@ -1,97 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import { isUUID } from '@authup/kit';
-import { useRequestQuery } from '@routup/basic/query';
-import type { Request, Response } from 'routup';
-import { send, useRequestParam } from 'routup';
-import {
-    applyQuery,
-} from 'typeorm-extension';
-import { NotFoundError } from '@ebec/http';
-import type { DataSource } from 'typeorm';
-import { BucketEntity } from '../../../../database/index.ts';
-
-export async function executeBucketRouteGetOneHandler(
-    req: Request,
-    res: Response,
-    dataSource: DataSource,
-) : Promise<any> {
-    const id = useRequestParam(req, 'id');
-
-    const repository = dataSource.getRepository(BucketEntity);
-    const query = repository.createQueryBuilder('bucket');
-
-    if (isUUID(id)) {
-        query.where('bucket.id = :id', { id });
-    } else {
-        query.where('bucket.name LIKE :name', { name: id });
-    }
-
-    applyQuery(query, useRequestQuery(req), {
-        defaultAlias: 'bucket',
-        fields: {
-            default: [
-                'id',
-                'name',
-                'region',
-                'created_at',
-                'updated_at',
-                'realm_id',
-                'actor_type',
-                'actor_id',
-            ],
-        },
-    });
-
-    const entity = await query.getOne();
-
-    if (!entity) {
-        throw new NotFoundError();
-    }
-
-    return send(res, entity);
-}
-
-export async function executeBucketRouteGetManyHandler(
-    req: Request,
-    res: Response,
-    dataSource: DataSource,
-) : Promise<any> {
-    const repository = dataSource.getRepository(BucketEntity);
-    const query = repository.createQueryBuilder('bucket');
-    query.groupBy('bucket.id');
-
-    const { pagination } = applyQuery(query, useRequestQuery(req), {
-        defaultAlias: 'bucket',
-        fields: {
-            default: [
-                'id',
-                'name',
-                'region',
-                'created_at',
-                'updated_at',
-                'realm_id',
-                'actor_id',
-                'actor_type',
-            ],
-        },
-        filters: { allowed: ['id', 'name', 'realm_id', 'actor_type', 'actor_id'] },
-        pagination: { maxLimit: 50 },
-        sort: { allowed: ['id', 'updated_at', 'created_at'] },
-    });
-
-    const [entities, total] = await query.getManyAndCount();
-
-    return send(res, {
-        data: entities,
-        meta: {
-            total,
-            ...pagination,
-        },
-    });
-}
