--- apps/server-telemetry/src/adapters/http/controllers/event/handlers/index.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2025.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export * from './create.ts';
-export * from './delete.ts';
-export * from './read.ts';
